+++ src/store/index.ts
@@ -1,14 +1,18 @@
-import { configureStore } from '@reduxjs/toolkit'
+import { combineReducers, configureStore } from '@reduxjs/toolkit'
+import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux'
+
+import PostsSlice from '@/store/slices/posts.slice'
 
 export const store = configureStore({
-  reducer: {},
+  reducer: combineReducers({
+    postsSlice: PostsSlice,
+  }),
 })
 
-// Infer the `RootState` and `AppDispatch` types from the store itself
 export type RootState = ReturnType<typeof store.getState>
-// Inferred type: {
-//   posts: PostsState,
-//   comments: CommentsState,
-//   users: UsersState
-// }
 export type AppDispatch = typeof store.dispatch
+
+export const useAppDispatch = () => useDispatch<AppDispatch>()
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
+
+export * from './slices/posts.slice'